refactor: common 폴더 분리 (#5)
