+++ src/main/java/com/kt/common/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.kt.common;
+package com.kt.common.response;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
