+++ src/main/java/com/kt/common/support/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.kt.common;
+package com.kt.common.support;
 
 import java.time.LocalDateTime;
