+++ src/main/java/com/kt/common/support/Preconditions.java
@@ -1,4 +1,7 @@
-package com.kt.common;
+package com.kt.common.support;
+
+import com.kt.common.exception.CustomException;
+import com.kt.common.exception.ErrorCode;
 
 public class Preconditions {
 	public static void validate(boolean expression, ErrorCode errorCode) {
