+++ src/main/java/com/kt/controller/category/CategoryController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.kt.common.ApiResult;
+import com.kt.common.response.ApiResult;
 import com.kt.dto.category.CategoryCreateRequest;
 import com.kt.service.category.CategoryService;
