+++ src/main/java/com/kt/controller/membership/AdminMembershipController.java
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.kt.common.ApiResult;
-import com.kt.common.Paging;
+import com.kt.common.response.ApiResult;
+import com.kt.common.request.Paging;
 import com.kt.dto.membership.MembershipCreateRequest;
 import com.kt.dto.membership.MembershipListResponse;
 import com.kt.dto.membership.MembershipUpdateRequest;
