+++ src/main/java/com/kt/controller/user/UserController.java
@@ -1,6 +1,6 @@
 package com.kt.controller.user;
 
-import com.kt.common.ApiResult;
+import com.kt.common.response.ApiResult;
 import com.kt.dto.user.UserCreateRequest;
 import com.kt.dto.user.UserLoginRequest;
 import com.kt.dto.user.UserLoginResponse;
