+++ src/main/java/com/kt/domain/category/Category.java
@@ -3,15 +3,9 @@ package com.kt.domain.category;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.apache.logging.log4j.util.Strings;
-
-import com.kt.common.ErrorCode;
-import com.kt.common.Preconditions;
 import com.kt.domain.product.Product;
 
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
