+++ src/main/java/com/kt/repository/category/CategoryRepository.java
@@ -2,8 +2,8 @@ package com.kt.repository.category;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.kt.common.CustomException;
-import com.kt.common.ErrorCode;
+import com.kt.common.exception.CustomException;
+import com.kt.common.exception.ErrorCode;
 import com.kt.domain.category.Category;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
