+++ src/main/java/com/kt/repository/product/ProductRepository.java
@@ -2,8 +2,8 @@ package com.kt.repository.product;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.kt.common.CustomException;
-import com.kt.common.ErrorCode;
+import com.kt.common.exception.CustomException;
+import com.kt.common.exception.ErrorCode;
 import com.kt.domain.product.Product;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
