+++ src/main/java/com/kt/service/review/ReviewService.java
@@ -6,8 +6,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.kt.common.ErrorCode;
-import com.kt.common.Preconditions;
+import com.kt.common.exception.ErrorCode;
+import com.kt.common.support.Preconditions;
 import com.kt.domain.review.Review;
 import com.kt.dto.review.ReviewCreateRequest;
 import com.kt.dto.review.ReviewListResponse;
